fixing review changes
