+++ systest/21million/bulk/run_test.go
@@ -1,9 +1,27 @@
+/*
+ * Copyright 2020 Dgraph Labs, Inc. and Contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package bulk
 
 import (
+	"os"
+	"path"
+
 	"github.com/dgraph-io/dgraph/systest/21million/common"
 	"github.com/dgraph-io/dgraph/testutil"
-	"os"
 
 	"testing"
 )
@@ -13,8 +31,8 @@ func TestQueries(t *testing.T) {
 }
 
 func TestMain(m *testing.M) {
-	schemaFile := os.Getenv("TEST_DATA_DIRECTORY") + "/21million.schema"
-	rdfFile := os.Getenv("TEST_DATA_DIRECTORY") + "/21million.rdf.gz"
+	schemaFile := path.Join(testutil.TestDataDirectory, "21million.schema")
+	rdfFile := path.Join(testutil.TestDataDirectory, "21million.rdf.gz")
 	if err := testutil.MakeDirEmpty([]string{"out/0", "out/1", "out/2"}); err != nil {
 		os.Exit(1)
 	}
