+++ systest/21million/live/run_test.go
@@ -1,8 +1,26 @@
+/*
+ * Copyright 2020 Dgraph Labs, Inc. and Contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package bulk
 
 import (
-	"github.com/dgraph-io/dgraph/testutil"
 	"os"
+	"path"
+
+	"github.com/dgraph-io/dgraph/testutil"
 
 	"github.com/dgraph-io/dgraph/systest/21million/common"
 
@@ -14,8 +32,8 @@ func TestQueries(t *testing.T) {
 }
 
 func TestMain(m *testing.M) {
-	schemaFile := os.Getenv("TEST_DATA_DIRECTORY") + "/21million.schema"
-	rdfFile := os.Getenv("TEST_DATA_DIRECTORY") + "/21million.rdf.gz"
+	schemaFile := path.Join(testutil.TestDataDirectory, "21million.schema")
+	rdfFile := path.Join(testutil.TestDataDirectory, "21million.rdf.gz")
 	if err := testutil.LiveLoad(testutil.LiveOpts{
 		Alpha:      testutil.ContainerAddr("alpha1", 9080),
 		Zero:       testutil.SockAddrZero,
