+++ systest/21million/ludicrous/run_test.go
@@ -1,21 +1,40 @@
+/*
+ * Copyright 2020 Dgraph Labs, Inc. and Contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package bulk
 
 import (
 	"fmt"
-	"github.com/dgraph-io/dgraph/testutil"
 	"os"
 	"os/exec"
+	"path"
+
+	"github.com/dgraph-io/dgraph/testutil"
 
 	"testing"
 )
+
 //
 //func TestQueries(t *testing.T) {
 //	t.Run("Run queries", common.TestQueriesFor21Million)
 //}
 
 func TestMain(m *testing.M) {
-	schemaFile := os.Getenv("TEST_DATA_DIRECTORY") + "/21million.schema"
-	rdfFile := os.Getenv("TEST_DATA_DIRECTORY") + "/21million.rdf.gz"
+	schemaFile := path.Join(testutil.TestDataDirectory, "21million.schema")
+	rdfFile := path.Join(testutil.TestDataDirectory, "21million.rdf.gz")
 
 	liveCmd := exec.Command(testutil.DgraphBinaryPath(), "live",
 		"--files", rdfFile,
