+++ systest/bulk_live/live/live_test.go
@@ -1,8 +1,25 @@
+/*
+ * Copyright 2020 Dgraph Labs, Inc. and Contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package live
 
 import (
-	"github.com/dgraph-io/dgraph/systest/bulk_live/common"
 	"testing"
+
+	"github.com/dgraph-io/dgraph/systest/bulk_live/common"
 )
 
 func TestLiveCases(t *testing.T) {
